backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
tree-sitter==0.23.0
tree-sitter-python==0.23.0
tree-sitter-javascript==0.23.0
neo4j==5.14.1
chromadb==0.5.23
sentence-transformers==3.0.0
gitpython==3.1.40
networkx==3.2.1
pydantic==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.6
python-dotenv==1.0.0
groq==0.11.0
httpx==0.27.0
